# packages/contracts/openapi-platform.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Impact control-plane API",
    "version": "1.2.0"
  },
  "paths": {
...
    "/v1/platform/tenants/{tenant_id}/access-bootstrap": {
      "parameters": [
        {
          "name": "tenant_id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "post": {
        "operationId": "request_initial_access",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "description": "Control-plane authority or nominated identity; never tenant role inheritance. Cookie mutations require Origin and X-CSRF-Token. Mutations require fresh MFA assurance. Activation requires an independent natural identity.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "request_id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "tenant_id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "revision_id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "operating_name": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 200,
                      "pattern": "\\S"
                    },
                    "owner_identity_id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "second_identity_id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "state": {
                      "enum": [
                        "Requested",
                        "Accepted",
                        "Applied",
                        "Rejected",
                        "Cancelled"
                      ]
                    },
                    "manifest": {
                      "type": "object",
                      "additionalProperties": false,
                      "properties": {
                        "version": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 80,
                          "pattern": "\\S"
                        },
                        "roles": {
                          "type": "object",
                          "additionalProperties": false,
                          "properties": {
                            "TENANT_ADMIN": {
                              "type": "array",
                              "items": {
                                "type": "string",
                                "minLength": 1,
                                "maxLength": 64,
                                "pattern": "\\S"
                              },
                              "uniqueItems": true,
                              "maxItems": 200
                            },
                            "AUTHOR": {
                              "type": "array",
                              "items": {
                                "type": "string",
                                "minLength": 1,
                                "maxLength": 64,
                                "pattern": "\\S"
                              },
                              "uniqueItems": true,
                              "maxItems": 200
                            },
                            "REVIEWER": {
                              "type": "array",
                              "items": {
                                "type": "string",
                                "minLength": 1,
                                "maxLength": 64,
                                "pattern": "\\S"
                              },
                              "uniqueItems": true,
                              "maxItems": 200
                            },
                            "PROGRAMME_MANAGER": {
                              "type": "array",
                              "items": {
                                "type": "string",
                                "minLength": 1,
                                "maxLength": 64,
                                "pattern": "\\S"
                              },
                              "uniqueItems": true,
                              "maxItems": 200
                            },
                            "ANALYST": {
                              "type": "array",
                              "items": {
                                "type": "string",
                                "minLength": 1,
                                "maxLength": 64,
                                "pattern": "\\S"
                              },
                              "uniqueItems": true,
                              "maxItems": 200
                            },
                            "EXTERNAL": {
                              "type": "array",
                              "items": {
                                "type": "string",
                                "minLength": 1,
                                "maxLength": 64,
                                "pattern": "\\S"
                              },
                              "uniqueItems": true,
                              "maxItems": 200
                            }
                          },
                          "required": [
                            "TENANT_ADMIN"
                          ]
                        }
                      },
                      "required": [
                        "version",
                        "roles"
                      ]
                    },
                    "profile_hash": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 64,
                      "pattern": "\\S"
                    },
                    "reason": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 1000,
                      "pattern": "\\S"
                    },
                    "expires_at": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "review_expires_at": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "scope_id": {
                      "anyOf": [
                        {
                          "type": "string",
                          "format": "uuid"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "second_membership_id": {
                      "anyOf": [
                        {
                          "type": "string",
                          "format": "uuid"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "operation_id": {
                      "type": "string",
                      "format": "uuid"
                    }
                  },
                  "required": [
                    "request_id",
                    "tenant_id",
                    "revision_id",
                    "operating_name",
                    "owner_identity_id",
                    "second_identity_id",
                    "state",
                    "manifest",
                    "profile_hash",
                    "reason",
                    "expires_at",
                    "review_expires_at",
                    "scope_id",
                    "second_membership_id",
                    "operation_id"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Structured domain error: code, message, retryable, correlation_id, permitted_actions, field_errors, optional reason_code."
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "operation_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "expected_revision": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "data": {
                    "type": "object",
                    "additionalProperties": false,
                    "properties": {
                      "second_identity_id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "profile_hash": {
                        "type": "string",
                        "pattern": "^[0-9a-f]{64}$"
                      },
                      "role_names": {
                        "type": "array",
                        "items": {
                          "enum": [
                            "AUTHOR",
                            "REVIEWER",
                            "PROGRAMME_MANAGER",
                            "ANALYST",
                            "EXTERNAL"
                          ]
                        },
                        "minItems": 1,
                        "maxItems": 5,
                        "uniqueItems": true
                      },
                      "expires_at": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "reason": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 1000,
                        "pattern": "\\S"
                      }
                    },
                    "required": [
                      "second_identity_id",
                      "profile_hash",
                      "role_names",
                      "expires_at",
                      "reason"
                    ]
                  }
                },
                "required": [
                  "operation_id",
                  "expected_revision",
                  "data"
                ]
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-impact_session"
      }
    }
  }
}